remove テストコードの余分な部分を削除しました.
