+++ Tests/TsumijiTests/TsumijiTests.swift
@@ -27,6 +27,15 @@ extension Color {
     static let defaultBackground = Color.clear
 }
 
+extension Attribute {
+    static let red: Self = .fontColor(.red)
+    static let big: Self = .fontSize(50)
+    static let impact: Self = .fontName("times")
+    static let blueBack: Self = .backgroundColor(.blue)
+    
+    static let bigRed: Self = .fontSize(50).fontColor(.red)
+}
+
 final class TsumijiTests: XCTestCase {
     
     let singleElementAttrCase = NSMutableAttributedString()
